pyproject.toml
[project]
name = "hack"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "arxiv>=2.2.0",
    "beautifulsoup4>=4.13.4",
    "fastmcp>=2.11.2",
    "httpx>=0.28.1",
    "markdownify>=1.1.0",
    "mcp>=1.12.4",
    "paho-mqtt>=2.1.0",
    "pillow>=11.3.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "readabilipy>=0.3.0",
]
